// app/(auth)/apps/chat/types.ts
export type ChatItemProps = {
  id: number;
  name?: string;
  image?: string;
  date?: string;
  status?: "sent" | "forwarded" | "read";
  is_archive?: boolean;
  last_message?: string;
  messages?: ChatMessageProps[];
  user_id: number;
  user: UserPropsTypes;
  type?: "DIRECT" | "GROUP"; // Chat type
};

export type ChatMessageProps = {
  id: number;
  content?: string;
  type?: string;
  own_message?: boolean;
  read?: boolean;
  data?: ChatMessageDataProps;
  createdAt?: string;
  updatedAt?: string;
  isEdited?: boolean;
};

export type ChatMessageDataProps = {
  file_name?: string;
  name?: string;
  cover?: string;
  path?: string;
  duration?: string;
  size?: string;
  images?: [];
  starred?: boolean;
};

export type UserPropsTypes = {
  id: number;
  name: string;
  avatar?: string;
  about?: string;
  phone?: string;
  country?: string;
  email?: string;
  gender?: string;
  website?: string;
  online_status?: "success" | "warning" | "danger";
  last_seen?: string;
  social_links?: {
    name?: string;
    url?: string;
  }[];
  medias?: {
    type?: string;
    path?: string;
  }[];
};

export type MediaListItemType = {
  type: string;
};

export type MessageStatusIconType = {
  status?: string;
};